feat 订单接口重构

+++ src/main/java/com/joe/util/mvc/ResponsePageEntity.java
@@ -17,7 +17,7 @@ public class ResponsePageEntity {
     /**
      * 记录总条数
      */
-    private int total;
+    private long total;
 
     /**
      * 数据
